+++ stash/BasicToken.sol
